Fix #8: Set *default-pathname-defaults* for cmucl.
Set `*default-pathname-defaults*` for cmucl in `gclload1.lsp` so that
it has the full path to `gclload.lsp`.  This is needed so that the
tests can be loaded from the subdirectories.  Without this, each load
file in the subdirectory is set to #P"" and when it tries to load a
test file, it can't because the setting of
`*default-pathname-defaults*` isn't pointing to the subdirectory but
rather the main directory since #P"" means current directory, and
cmucl hasn't changed it's current directory.

This fix only applies to cmucl and no other lisps.