Extensions and fixes to the random tester
Add some macros for common cases.  Using macros makes
these easier to generate and prune.

Fix some pruning problems (mostly forgetting to call
the try function.)

Extended the range of integers to catch some bignum issues.

Added special pruning for boolean expressions and for top level
expressions (where we don't care if we preserve the type of the
expression.)

Add the possibility of passing boolean variables in as parameters
to the random function.