Add testing mode for pruner
The random form generator can be told to produce
forms that signal an error at compile time.  The
pruner should reduce forms with these to bare
forms containing just this.  Anything beyond that
is fluff that the pruner should have been able to get
rid of.

Fix a bunch of deficiencies in the pruner revealed
by this testing.